backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.2
vaderSentiment==3.3.2
textblob==0.17.1
transformers==4.36.0
torch>=2.6.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
asyncio==3.4.3
schedule==1.2.0
aiofiles==23.2.1

# Reddit API client
praw==7.7.0
